Further share code between eligibility and subcategory components. Refactor and consolidate styles.
